requirements.txt
# FastAPI
fastapi[standard]==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
greenlet==3.1.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Utils
python-multipart==0.0.20
python-dotenv==1.0.1
httpx[http2]==0.28.1

# AI - OpenAI
openai>=1.58.0

# AI - Gemini TTS
google-genai>=1.0.0

# Image processing
Pillow>=10.0.0

# Sync DB engine (for AI memory manager)
psycopg2-binary==2.9.10
